refactor: reorg HotkeyManager.lua to Player/
