+++ TTS_xwing/src/Player/HotkeyManager.lua
